zstd: Use int type for return type of public functions

+++ include/fluent-bit/flb_zstd.h
@@ -26,8 +26,8 @@
 
 struct flb_decompression_context;
 
-size_t flb_zstd_compress(void *in_data, size_t in_len, void **out_data, size_t *out_len);
-size_t flb_zstd_uncompress(void *in_data, size_t in_len, void **out_data, size_t *out_len);
+int flb_zstd_compress(void *in_data, size_t in_len, void **out_data, size_t *out_len);
+int flb_zstd_uncompress(void *in_data, size_t in_len, void **out_data, size_t *out_len);
 
 int flb_zstd_decompressor_dispatch(struct flb_decompression_context *context,
                                    void *output_buffer,
